+++ external-crates/move/crates/move-compiler/src/linters/mod.rs
@@ -5,9 +5,9 @@ use move_symbol_pool::Symbol;
 
 use crate::{
     command_line::compiler::Visitor, diagnostics::codes::WarningFilter,
-    linters::shift_overflow::ShiftOperationOverflow, typing::visitor::TypingVisitor,
+    linters::unnecessary_while_loop::WhileTrueToLoop, typing::visitor::TypingVisitor,
 };
-pub mod shift_overflow;
+pub mod unnecessary_while_loop;
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum LintLevel {
     // No linters
@@ -20,12 +20,12 @@ pub enum LintLevel {
 
 pub const ALLOW_ATTR_CATEGORY: &str = "lint";
 pub const LINT_WARNING_PREFIX: &str = "Lint ";
-pub const SHILF_OVERFLOW_FILTER_NAME: &str = "shift_overflow";
+pub const WHILE_TRUE_TO_LOOP_FILTER_NAME: &str = "unnecessary_while_loop";
 
 pub const LINTER_DEFAULT_DIAG_CODE: u8 = 1;
 
 pub enum LinterDiagCategory {
-    ShiftOperationOverflow,
+    WhileTrueToLoop,
 }
 
 pub fn known_filters() -> (Option<Symbol>, Vec<WarningFilter>) {
@@ -33,9 +33,9 @@ pub fn known_filters() -> (Option<Symbol>, Vec<WarningFilter>) {
         Some(ALLOW_ATTR_CATEGORY.into()),
         vec![WarningFilter::code(
             Some(LINT_WARNING_PREFIX),
-            LinterDiagCategory::ShiftOperationOverflow as u8,
+            LinterDiagCategory::WhileTrueToLoop as u8,
             LINTER_DEFAULT_DIAG_CODE,
-            Some(SHILF_OVERFLOW_FILTER_NAME),
+            Some(WHILE_TRUE_TO_LOOP_FILTER_NAME),
         )],
     )
 }
@@ -44,8 +44,8 @@ pub fn linter_visitors(level: LintLevel) -> Vec<Visitor> {
     match level {
         LintLevel::None => vec![],
         LintLevel::Default | LintLevel::All => {
-            vec![shift_overflow::ShiftOperationOverflow::visitor(
-                ShiftOperationOverflow,
+            vec![unnecessary_while_loop::WhileTrueToLoop::visitor(
+                WhileTrueToLoop,
             )]
         }
     }
