+++ external-crates/move/crates/move-compiler/tests/custom_rules/unnecessary_while_loop.move
@@ -1,0 +1,12 @@
+module 0x42::M {
+
+    public fun finite_loop() {
+        let counter = 0;
+        while (true) {
+            if(counter == 10) {
+                break
+            };
+            counter = counter + 1;
+        }
+    }
+}
